Added ttl for cookies and refactored
