+++ lib/session.ts
@@ -6,22 +6,17 @@ import { withIronSession } from "next-iron-session"
 import "./queue-processor"
 import "./exit-strategies"
 import "./watchers"
-import { toIst } from "./utils"
+import { secondsTill7 } from "./utils"
 
 export default function withSession(handler) {
-  const nowIst = toIst(new Date())
-  const next8AmIst = nowIst.hour() >= 8
-    ? nowIst.add(1, "day").startOf("day").hour(8)
-    : nowIst.startOf("day").hour(8)
-  const secondsTill8: number = next8AmIst.diff(nowIst, "second")
   return withIronSession(handler, {
     password: process.env.SECRET_COOKIE_PASSWORD!,
     cookieName: "khaching/kite/session",
+    ttl: secondsTill7(),
     cookieOptions: {
       // the next line allows to use the session in non-https environments like
       // Next.js dev mode (http://localhost:3000)
       secure: process.env.NODE_ENV === "production",
-      maxAge: secondsTill8,
     },
   })
 }
