+++ package.json
@@ -36,7 +36,6 @@
     "axios": "^0.21.1",
     "bluebird": "^3.7.2",
     "bullmq": "^5.73.3",
-    "csvtojson": "^2.0.10",
     "date-fns": "^2.19.0",
     "dayjs": "^1.10.4",
     "dotenv": "^10.0.0",
