+++ pages/api/prices.js
@@ -1,5 +1,5 @@
 import withSession from "../../lib/session"
-import { getTradingSymbolsByOptionPrice } from "../../lib/utils"
+import { getTradingSymbolsByOptionPrice } from "../../lib/kiteUtils"
 
 export default withSession(async (req, res) => {
   const user = req.session.get("user")
